renamed eigencuda to cudapipeline
